fix: Update migration guide, import docs, and task tests (#3933)
- Update migration guide for `storage_aws_external_id` (references
#3924)
- Add `ancient` version clarification to the migration guide
- Clarify import docs for `database`, `secondary_database`, and `schema`
(references #3923)
- Move task integration tests and one task acceptance test to account
level tests
