+++ pkg/testacc/resource_task_account_level_acceptance_test.go
@@ -1,0 +1,321 @@
+//go:build account_level_tests
+
+package testacc
+
+import (
+	"testing"
+
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/assert/objectparametersassert"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/assert/resourceparametersassert"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config/model"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/helpers"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/provider/resources"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/tfversion"
+)
+
+// temporarily moved to account level tests due to STATEMENT_TIMEOUT_IN_SECONDS being set on warehouse level and messing with the results
+func TestAcc_Task_AllParameters(t *testing.T) {
+	id := testClient().Ids.RandomSchemaObjectIdentifier()
+	statement := "SELECT 1"
+
+	configModel := model.TaskWithId("test", id, true, statement).
+		WithScheduleMinutes(5)
+	configModelWithAllParametersSet := model.TaskWithId("test", id, true, statement).
+		WithScheduleMinutes(5).
+		WithSuspendTaskAfterNumFailures(15).
+		WithTaskAutoRetryAttempts(15).
+		WithUserTaskManagedInitialWarehouseSizeEnum(sdk.WarehouseSizeXSmall).
+		WithUserTaskMinimumTriggerIntervalInSeconds(30).
+		WithUserTaskTimeoutMs(1000).
+		WithAbortDetachedQuery(true).
+		WithAutocommit(false).
+		WithBinaryInputFormatEnum(sdk.BinaryInputFormatUTF8).
+		WithBinaryOutputFormatEnum(sdk.BinaryOutputFormatBase64).
+		WithClientMemoryLimit(1024).
+		WithClientMetadataRequestUseConnectionCtx(true).
+		WithClientPrefetchThreads(2).
+		WithClientResultChunkSize(48).
+		WithClientResultColumnCaseInsensitive(true).
+		WithClientSessionKeepAlive(true).
+		WithClientSessionKeepAliveHeartbeatFrequency(2400).
+		WithClientTimestampTypeMappingEnum(sdk.ClientTimestampTypeMappingNtz).
+		WithDateInputFormat("YYYY-MM-DD").
+		WithDateOutputFormat("YY-MM-DD").
+		WithEnableUnloadPhysicalTypeOptimization(false).
+		WithErrorOnNondeterministicMerge(false).
+		WithErrorOnNondeterministicUpdate(true).
+		WithGeographyOutputFormatEnum(sdk.GeographyOutputFormatWKB).
+		WithGeometryOutputFormatEnum(sdk.GeometryOutputFormatWKB).
+		WithJdbcUseSessionTimezone(false).
+		WithJsonIndent(4).
+		WithLockTimeout(21222).
+		WithLogLevelEnum(sdk.LogLevelError).
+		WithMultiStatementCount(0).
+		WithNoorderSequenceAsDefault(false).
+		WithOdbcTreatDecimalAsInt(true).
+		WithQueryTag("some_tag").
+		WithQuotedIdentifiersIgnoreCase(true).
+		WithRowsPerResultset(2).
+		WithS3StageVpceDnsName("vpce-id.s3.region.vpce.amazonaws.com").
+		WithSearchPath("$public, $current").
+		WithStatementQueuedTimeoutInSeconds(10).
+		WithStatementTimeoutInSeconds(10).
+		WithStrictJsonOutput(true).
+		WithTimestampDayIsAlways24h(true).
+		WithTimestampInputFormat("YYYY-MM-DD").
+		WithTimestampLtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+		WithTimestampNtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+		WithTimestampOutputFormat("YYYY-MM-DD HH24:MI:SS").
+		WithTimestampTypeMappingEnum(sdk.TimestampTypeMappingLtz).
+		WithTimestampTzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+		WithTimezone("Europe/Warsaw").
+		WithTimeInputFormat("HH24:MI").
+		WithTimeOutputFormat("HH24:MI").
+		WithTraceLevelEnum(sdk.TraceLevelPropagate).
+		WithTransactionAbortOnError(true).
+		WithTransactionDefaultIsolationLevelEnum(sdk.TransactionDefaultIsolationLevelReadCommitted).
+		WithTwoDigitCenturyStart(1980).
+		WithUnsupportedDdlActionEnum(sdk.UnsupportedDDLActionFail).
+		WithUseCachedResult(false).
+		WithWeekOfYearPolicy(1).
+		WithWeekStart(1)
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: TestAccProtoV6ProviderFactories,
+		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
+			tfversion.RequireAbove(tfversion.Version1_5_0),
+		},
+		PreCheck:     func() { TestAccPreCheck(t) },
+		CheckDestroy: CheckDestroy(t, resources.User),
+		Steps: []resource.TestStep{
+			// create with default values for all the parameters
+			{
+				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
+				ConfigVariables: config.ConfigVariablesFromModel(t, configModel),
+				Check: assertThat(t,
+					objectparametersassert.TaskParameters(t, id).
+						HasAllDefaults().
+						HasAllDefaultsExplicit(),
+					resourceparametersassert.TaskResourceParameters(t, configModel.ResourceReference()).
+						HasAllDefaults(),
+				),
+			},
+			// import when no parameter set
+			{
+				ResourceName:    configModel.ResourceReference(),
+				ImportState:     true,
+				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
+				ConfigVariables: config.ConfigVariablesFromModel(t, configModel),
+				ImportStateCheck: assertThatImport(t,
+					resourceparametersassert.ImportedTaskResourceParameters(t, helpers.EncodeResourceIdentifier(id)).
+						HasAllDefaults(),
+				),
+			},
+			// set all parameters
+			{
+				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
+				ConfigVariables: config.ConfigVariablesFromModel(t, configModelWithAllParametersSet),
+				Check: assertThat(t,
+					objectparametersassert.TaskParameters(t, id).
+						HasSuspendTaskAfterNumFailures(15).
+						HasTaskAutoRetryAttempts(15).
+						HasUserTaskManagedInitialWarehouseSize(sdk.WarehouseSizeXSmall).
+						HasUserTaskMinimumTriggerIntervalInSeconds(30).
+						HasUserTaskTimeoutMs(1000).
+						HasAbortDetachedQuery(true).
+						HasAutocommit(false).
+						HasBinaryInputFormat(sdk.BinaryInputFormatUTF8).
+						HasBinaryOutputFormat(sdk.BinaryOutputFormatBase64).
+						HasClientMemoryLimit(1024).
+						HasClientMetadataRequestUseConnectionCtx(true).
+						HasClientPrefetchThreads(2).
+						HasClientResultChunkSize(48).
+						HasClientResultColumnCaseInsensitive(true).
+						HasClientSessionKeepAlive(true).
+						HasClientSessionKeepAliveHeartbeatFrequency(2400).
+						HasClientTimestampTypeMapping(sdk.ClientTimestampTypeMappingNtz).
+						HasDateInputFormat("YYYY-MM-DD").
+						HasDateOutputFormat("YY-MM-DD").
+						HasEnableUnloadPhysicalTypeOptimization(false).
+						HasErrorOnNondeterministicMerge(false).
+						HasErrorOnNondeterministicUpdate(true).
+						HasGeographyOutputFormat(sdk.GeographyOutputFormatWKB).
+						HasGeometryOutputFormat(sdk.GeometryOutputFormatWKB).
+						HasJdbcUseSessionTimezone(false).
+						HasJsonIndent(4).
+						HasLockTimeout(21222).
+						HasLogLevel(sdk.LogLevelError).
+						HasMultiStatementCount(0).
+						HasNoorderSequenceAsDefault(false).
+						HasOdbcTreatDecimalAsInt(true).
+						HasQueryTag("some_tag").
+						HasQuotedIdentifiersIgnoreCase(true).
+						HasRowsPerResultset(2).
+						HasS3StageVpceDnsName("vpce-id.s3.region.vpce.amazonaws.com").
+						HasSearchPath("$public, $current").
+						HasStatementQueuedTimeoutInSeconds(10).
+						HasStatementTimeoutInSeconds(10).
+						HasStrictJsonOutput(true).
+						HasTimestampDayIsAlways24h(true).
+						HasTimestampInputFormat("YYYY-MM-DD").
+						HasTimestampLtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampNtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampTypeMapping(sdk.TimestampTypeMappingLtz).
+						HasTimestampTzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimezone("Europe/Warsaw").
+						HasTimeInputFormat("HH24:MI").
+						HasTimeOutputFormat("HH24:MI").
+						HasTraceLevel(sdk.TraceLevelPropagate).
+						HasTransactionAbortOnError(true).
+						HasTransactionDefaultIsolationLevel(sdk.TransactionDefaultIsolationLevelReadCommitted).
+						HasTwoDigitCenturyStart(1980).
+						HasUnsupportedDdlAction(sdk.UnsupportedDDLActionFail).
+						HasUseCachedResult(false).
+						HasWeekOfYearPolicy(1).
+						HasWeekStart(1),
+					resourceparametersassert.TaskResourceParameters(t, configModelWithAllParametersSet.ResourceReference()).
+						HasSuspendTaskAfterNumFailures(15).
+						HasTaskAutoRetryAttempts(15).
+						HasUserTaskManagedInitialWarehouseSize(sdk.WarehouseSizeXSmall).
+						HasUserTaskMinimumTriggerIntervalInSeconds(30).
+						HasUserTaskTimeoutMs(1000).
+						HasAbortDetachedQuery(true).
+						HasAutocommit(false).
+						HasBinaryInputFormat(sdk.BinaryInputFormatUTF8).
+						HasBinaryOutputFormat(sdk.BinaryOutputFormatBase64).
+						HasClientMemoryLimit(1024).
+						HasClientMetadataRequestUseConnectionCtx(true).
+						HasClientPrefetchThreads(2).
+						HasClientResultChunkSize(48).
+						HasClientResultColumnCaseInsensitive(true).
+						HasClientSessionKeepAlive(true).
+						HasClientSessionKeepAliveHeartbeatFrequency(2400).
+						HasClientTimestampTypeMapping(sdk.ClientTimestampTypeMappingNtz).
+						HasDateInputFormat("YYYY-MM-DD").
+						HasDateOutputFormat("YY-MM-DD").
+						HasEnableUnloadPhysicalTypeOptimization(false).
+						HasErrorOnNondeterministicMerge(false).
+						HasErrorOnNondeterministicUpdate(true).
+						HasGeographyOutputFormat(sdk.GeographyOutputFormatWKB).
+						HasGeometryOutputFormat(sdk.GeometryOutputFormatWKB).
+						HasJdbcUseSessionTimezone(false).
+						HasJsonIndent(4).
+						HasLockTimeout(21222).
+						HasLogLevel(sdk.LogLevelError).
+						HasMultiStatementCount(0).
+						HasNoorderSequenceAsDefault(false).
+						HasOdbcTreatDecimalAsInt(true).
+						HasQueryTag("some_tag").
+						HasQuotedIdentifiersIgnoreCase(true).
+						HasRowsPerResultset(2).
+						HasS3StageVpceDnsName("vpce-id.s3.region.vpce.amazonaws.com").
+						HasSearchPath("$public, $current").
+						HasStatementQueuedTimeoutInSeconds(10).
+						HasStatementTimeoutInSeconds(10).
+						HasStrictJsonOutput(true).
+						HasTimestampDayIsAlways24h(true).
+						HasTimestampInputFormat("YYYY-MM-DD").
+						HasTimestampLtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampNtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampTypeMapping(sdk.TimestampTypeMappingLtz).
+						HasTimestampTzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimezone("Europe/Warsaw").
+						HasTimeInputFormat("HH24:MI").
+						HasTimeOutputFormat("HH24:MI").
+						HasTraceLevel(sdk.TraceLevelPropagate).
+						HasTransactionAbortOnError(true).
+						HasTransactionDefaultIsolationLevel(sdk.TransactionDefaultIsolationLevelReadCommitted).
+						HasTwoDigitCenturyStart(1980).
+						HasUnsupportedDdlAction(sdk.UnsupportedDDLActionFail).
+						HasUseCachedResult(false).
+						HasWeekOfYearPolicy(1).
+						HasWeekStart(1),
+				),
+			},
+			// import when all parameters set
+			{
+				ResourceName:    configModelWithAllParametersSet.ResourceReference(),
+				ImportState:     true,
+				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
+				ConfigVariables: config.ConfigVariablesFromModel(t, configModelWithAllParametersSet),
+				ImportStateCheck: assertThatImport(t,
+					resourceparametersassert.ImportedTaskResourceParameters(t, helpers.EncodeResourceIdentifier(id)).
+						HasSuspendTaskAfterNumFailures(15).
+						HasTaskAutoRetryAttempts(15).
+						HasUserTaskManagedInitialWarehouseSize(sdk.WarehouseSizeXSmall).
+						HasUserTaskMinimumTriggerIntervalInSeconds(30).
+						HasUserTaskTimeoutMs(1000).
+						HasAbortDetachedQuery(true).
+						HasAutocommit(false).
+						HasBinaryInputFormat(sdk.BinaryInputFormatUTF8).
+						HasBinaryOutputFormat(sdk.BinaryOutputFormatBase64).
+						HasClientMemoryLimit(1024).
+						HasClientMetadataRequestUseConnectionCtx(true).
+						HasClientPrefetchThreads(2).
+						HasClientResultChunkSize(48).
+						HasClientResultColumnCaseInsensitive(true).
+						HasClientSessionKeepAlive(true).
+						HasClientSessionKeepAliveHeartbeatFrequency(2400).
+						HasClientTimestampTypeMapping(sdk.ClientTimestampTypeMappingNtz).
+						HasDateInputFormat("YYYY-MM-DD").
+						HasDateOutputFormat("YY-MM-DD").
+						HasEnableUnloadPhysicalTypeOptimization(false).
+						HasErrorOnNondeterministicMerge(false).
+						HasErrorOnNondeterministicUpdate(true).
+						HasGeographyOutputFormat(sdk.GeographyOutputFormatWKB).
+						HasGeometryOutputFormat(sdk.GeometryOutputFormatWKB).
+						HasJdbcUseSessionTimezone(false).
+						HasJsonIndent(4).
+						HasLockTimeout(21222).
+						HasLogLevel(sdk.LogLevelError).
+						HasMultiStatementCount(0).
+						HasNoorderSequenceAsDefault(false).
+						HasOdbcTreatDecimalAsInt(true).
+						HasQueryTag("some_tag").
+						HasQuotedIdentifiersIgnoreCase(true).
+						HasRowsPerResultset(2).
+						HasS3StageVpceDnsName("vpce-id.s3.region.vpce.amazonaws.com").
+						HasSearchPath("$public, $current").
+						HasStatementQueuedTimeoutInSeconds(10).
+						HasStatementTimeoutInSeconds(10).
+						HasStrictJsonOutput(true).
+						HasTimestampDayIsAlways24h(true).
+						HasTimestampInputFormat("YYYY-MM-DD").
+						HasTimestampLtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampNtzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimestampTypeMapping(sdk.TimestampTypeMappingLtz).
+						HasTimestampTzOutputFormat("YYYY-MM-DD HH24:MI:SS").
+						HasTimezone("Europe/Warsaw").
+						HasTimeInputFormat("HH24:MI").
+						HasTimeOutputFormat("HH24:MI").
+						HasTraceLevel(sdk.TraceLevelPropagate).
+						HasTransactionAbortOnError(true).
+						HasTransactionDefaultIsolationLevel(sdk.TransactionDefaultIsolationLevelReadCommitted).
+						HasTwoDigitCenturyStart(1980).
+						HasUnsupportedDdlAction(sdk.UnsupportedDDLActionFail).
+						HasUseCachedResult(false).
+						HasWeekOfYearPolicy(1).
+						HasWeekStart(1),
+				),
+			},
+			// unset all the parameters
+			{
+				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
+				ConfigVariables: config.ConfigVariablesFromModel(t, configModel),
+				Check: assertThat(t,
+					objectparametersassert.TaskParameters(t, id).
+						HasAllDefaults().
+						HasAllDefaultsExplicit(),
+					resourceparametersassert.TaskResourceParameters(t, configModel.ResourceReference()).
+						HasAllDefaults(),
+				),
+			},
+		},
+	})
+}
