+++ templates/resources/schema.md.tmpl
@@ -37,6 +37,9 @@ description: |-
 
 ## Import
 
+Make sure that `is_transient` and `with_managed_access` are set in your config explicitly before import to avoid diff and recreation
+(reasons are presented in [this migration guide entry](https://github.com/snowflakedb/terraform-provider-snowflake/blob/main/MIGRATION_GUIDE.md#behavior-change-boolean-type-changes-1)).
+
 Import is supported using the following syntax:
 
 {{ codefile "shell" (printf "examples/resources/%s/import.sh" .Name)}}
